# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Fidelito API
  version: 2.0.0
  description: |
    Digital loyalty card platform for Tunisian businesses.
    Clients display dynamic TOTP QR codes; cashiers scan to stamp loyalty cards.
    Businesses manage staff, rewards, and analytics. Admins manage merchants.
servers:
  - url: http://localhost:8787
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/client/qr/provision:
    post:
      summary: Provision TOTP secret and return QR
      description: Generates a TOTP secret for the authenticated client. Call once.
      responses:
        '200':
          description: QR payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret: { type: string, description: Base32 TOTP secret }
                  totp: { type: string, description: Current 6-digit TOTP code }
                  payload: { type: string, description: JSON QR payload }
                  qrDataUri: { type: string }
                  refreshInterval: { type: integer }
        '409': { description: Already provisioned }
  /api/client/qr/current:
    get:
      summary: Get current TOTP QR code
      description: Returns the current TOTP code and QR payload. Refreshes every 30s.
      responses:
        '200':
          description: Current QR
          content:
            application/json:
              schema:
                type: object
                properties:
                  totp: { type: string }
                  payload: { type: string }
                  qrDataUri: { type: string }
                  refreshInterval: { type: integer }
                  remainingSeconds: { type: integer }
        '400': { description: Not provisioned }
...
  /api/client/cards/{merchantId}:
    get:
      summary: Get card detail with transaction history
      parameters:
        - { name: merchantId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Card + transactions
        '404': { description: Card not found }
...